Added SSL support and fixed a bug in the destroy method that incorrectly returned an exception.
* README
Updated all examples to version 2.7 because of some fixes.

* lib\document.php
curlRequest: Creates a default CURL request.  If the request is SSL it turns on the proper options.
destroy: Now uses curlRequest.  No longer uses return transfer because this caused a failure code to be returned.
findById, upload: Now uses curlRequest.

* lib\service.php
Added documentation for making SSL requests.  Updated version to 1.1.0.

* test\destroy.php
Fixed the test so that it correctly returns if the request was successful.